requirements.txt
argon2-cffi==19.2.0
astroid==2.3.3
beautifulsoup4==4.8.0
blinker==1.4
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
django-mobile-app-distribution==0.5
edx-dl==0.1.10
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.2
opencv-python==4.1.2.30
passlib==1.7.1
Pillow==6.2.1
pycparser==2.19
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.12
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.9
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
wtf==0.1
WTForms==2.2.1
youtube-dl==2019.11.5
